+++ src/Phan/Language/Element/AddressableElement.php
@@ -112,7 +112,7 @@ abstract class AddressableElement extends TypedElement implements AddressableEle
      * @return bool true if this element's visibility
      *                   is strictly more visible than $other (public > protected > private)
      */
-    public function isStrictlyMoreVisibileThan(AddressableElementInterface $other) : bool
+    public function isStrictlyMoreVisibleThan(AddressableElementInterface $other) : bool
     {
         if ($this->isPrivate()) {
             return false;