+++ src/Phan/Language/Element/AddressableElementInterface.php
@@ -31,7 +31,7 @@ interface AddressableElementInterface extends TypedElementInterface
      * @return bool true if this element's visibility
      *                   is strictly more visible than $other (public > protected > private)
      */
-    public function isStrictlyMoreVisibileThan(AddressableElementInterface $other) : bool;
+    public function isStrictlyMoreVisibleThan(AddressableElementInterface $other) : bool;
 
     /**
      * @return bool